Added NavigationRibbon and NavigationRibbonItem with Initial Styles
Began to implement the NavigationRibbon control intended to replace the NavigationView in ShellPage.

The purpose of NavigationRibbon is to function as our top-level application navigation as well as providing a clean and consistent way for us to present contextual commands and options to the user for a given view/context.

The initial styling has been added to the `Themes/default.xaml` resource dictionary.